refactor: repository 목적에 따른 클래스 분리
