+++ src/main/java/cluverse/member/repository/MemberQueryRepository.java
@@ -1,26 +1,18 @@
 package cluverse.member.repository;
 
-import cluverse.member.domain.Member;
-import cluverse.member.domain.MemberMajor;
-import cluverse.member.domain.OAuthProvider;
-import cluverse.post.domain.PostStatus;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
 
 import static cluverse.member.domain.QBlock.block;
 import static cluverse.member.domain.QFollow.follow;
 import static cluverse.member.domain.QMember.member;
-import static cluverse.member.domain.QMemberAuth.memberAuth;
 import static cluverse.member.domain.QMemberMajor.memberMajor;
 import static cluverse.member.domain.QMemberProfile.memberProfile;
-import static cluverse.member.domain.QSocialAccount.socialAccount;
 import static cluverse.major.domain.QMajor.major;
-import static cluverse.post.domain.QPost.post;
 import static cluverse.interest.domain.QInterest.interest;
 import static cluverse.university.domain.QUniversity.university;
 
@@ -29,39 +21,6 @@ import static cluverse.university.domain.QUniversity.university;
 public class MemberQueryRepository {
 
     private final JPAQueryFactory queryFactory;
-
-    public boolean existsByEmail(String email) {
-        return queryFactory.selectFrom(memberAuth)
-                .where(memberAuth.email.eq(email))
-                .fetchFirst() != null;
-    }
-
-    public Optional<Member> findByEmail(String email) {
-        return Optional.ofNullable(
-                queryFactory.selectFrom(member)
-                        .join(member.memberAuth, memberAuth).fetchJoin()
-                        .where(memberAuth.email.eq(email))
-                        .fetchOne()
-        );
-    }
-
-    public Optional<Member> findBySocialAccount(OAuthProvider provider, String providerUserId) {
-        return Optional.ofNullable(
-                queryFactory.selectFrom(member)
-                        .join(member.socialAccounts, socialAccount).fetchJoin()
-                        .where(
-                                socialAccount.provider.eq(provider),
-                                socialAccount.providerUserId.eq(providerUserId)
-                        )
-                        .fetchOne()
-        );
-    }
-
-    public List<MemberMajor> findMajorsByMemberId(Long memberId) {
-        return queryFactory.selectFrom(memberMajor)
-                .where(memberMajor.member.id.eq(memberId))
-                .fetch();
-    }
 
     public List<MemberMajorDetailDto> findMajorDetailsByMemberId(Long memberId) {
         cluverse.major.domain.QMajor parentMajor = new cluverse.major.domain.QMajor("parentMajor");
@@ -234,17 +193,6 @@ public class MemberQueryRepository {
                         tuple.get(interest.category)
                 ))
                 .orElse(null);
-    }
-
-    public long countActivePostsByMemberId(Long memberId) {
-        Long count = queryFactory.select(post.count())
-                .from(post)
-                .where(
-                        post.memberId.eq(memberId),
-                        post.status.eq(PostStatus.ACTIVE)
-                )
-                .fetchOne();
-        return count == null ? 0L : count;
     }
 
     public record MemberMajorDetailDto(
