+++ src/main/java/cluverse/member/repository/MemberRepository.java
@@ -3,7 +3,7 @@ package cluverse.member.repository;
 import cluverse.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepository extends JpaRepository<Member, Long> {
+public interface MemberRepository extends JpaRepository<Member, Long>, MemberRepositoryCustom {
 
     boolean existsByNickname(String nickname);
 }
